Implemented shader metadata generation. Also implemented shader hashing to prevent rebuilds for shaders that haven't been changed since last build. It's still missing one crucial step with include files, which happens as follows. If a glsl (include file) is edited, the build process won't recognize that every shader that includes this newly-edited glsl file must also be indirectly rebuilt because of the include.
Separately, worked on reading in the generated shader metadata in the library code to utilize it when building the set layouts for a particular pass. This also needs more work before it's possible to programatically generate the SetLayoutSummary's, but it's getting there. In order to read in the generated checksum.json file, I added a new dependency (nlohmann_json) for reading in JSON files

+++ src/descriptors/set_layout/set_layout_summary.cpp
@@ -143,7 +143,7 @@ namespace TANG
 		for (auto& binding : bindings)
 		{
 			//pack the binding data into a single int64, including the set. Not fully correct but it's ok
-			size_t bindingHash = binding.binding | binding.descriptorType << 8 | binding.descriptorCount << 16 | binding.stageFlags << 24 | set << 32;
+			size_t bindingHash = binding.binding | binding.descriptorType << 8 | binding.descriptorCount << 16 | binding.stageFlags << 24 | static_cast<uint64_t>(set) << 32;
 
 			//shuffle the packed binding data and xor it with the main hash
 			hashResult ^= std::hash<size_t>()(bindingHash);
